Update package-lock.json and Docker related files
The package-lock.json file was updated to include the '@sirherobrine23/http_server' package with version 3.4.3 and GPL-2.0 license.

In Docker related files, the 'tags/list' path is now used instead of 'manifests/latest' in the 'docker/src/auth.ts' file. Additionally, a typo was fixed in the 'docker/src/registry.ts' file where 'manifet' was corrected to 'manifest'.

Signed-off-by: Matheus Sampaio Queiroga <srherobrine20@gmail.com>